requirements.txt
streamlit>=1.28.0
firebase-admin>=6.2.0
#stripe>=5.5.0
requests>=2.31.0
Pillow>=10.0.0
openai>=1.30.1
python-dotenv>=1.0.1
paypalrestsdk>=1.13.3 
